[core] support optional_top_level_keys in log declarations
The goal of optional_top_level_keys is to provide a mechanism for a more flexible schema.
If the key(s) described in optional_top_level_keys exist in the incoming record, they are simply
type checked.  If they do not exist, a default value is created in place to avoid
rules calling keys that do not exist.