[lambda] add a second layer of backoff to firehose
Currently, Firehose will not throw an exception if any of the records failed in the PutRecordBatch request.

It will however return details in the response of the number of failed records.  This change allows for clearer diagnosis of Firehose operations.

The logger message levels are temporary, I can move them back to debug after verifying backoff is working as expected.