Fix Github output mock credentials in `alert` testing
Extend the AlertProcessorTester.setup_outputs method to provide mock credentials
for the Github output and fix runtime exceptions caused by testing the alert
processor for rules that specify the Github output.

Extend the instructions in `docs/source/outputs.rst` to mention that this test
method should be updated for each new output that is being written.

Refs: https://github.com/airbnb/streamalert/pull/601