[core] fix process_subkeys nested key bug
Due to loose schemas, we implemented a required subkeys check in rules.  This subkey check ensures that the subkey both exists in the record, and has a Truthy value.  This same check is not applied a subkeys keys (aka nested keys).  Because of this, if you write rules based on a nested key, there's a chance that its value could be Falsey, which could lead to KeyErrors or.

This change makes sure that both subkeys and nested keys have valid Truthy values.