[testing] Update the alert merger test to disregard call order.
The `mock_logger.assert_has_calls(...)` asserts that the calls provided 
are executed in the order provided. However, in Python 3, the order of 
calls is different. The `any_oder=True` parameter is provided to the 
assertion to disregard this order.

Upon deeper analysis of the test semantics, it is testing to ensure that 
merging, deletion, and dispatching occur appropriately. The 
implementation semantics appear to be unchanged correct according to the 
test, and the only difference between the Python 2 and Python 3 versions 
is execution order.