[testing] Update the DuoApp test to properly patch abstract methods.
In Python 3 the `__abstractmethods__` variable contains all abstract 
methods which must be implemented in classes deriving from it. This 
causes problems when attempting to test the `DuoApp` class, since it 
does not implement the `_type` method.

Furthermore, the `patch.object` decorator will patch instances of a 
particular object in all methods matching the specified method name test 
prefix (e.g. `test_some_method` where the test prefix is `test_*`). This 
causes issues because the `setup` function does not match the test 
prefix, which results in an error when attempting to create an instance 
of `DuoApp` before each test.

This commit fixes this issue by patching all methods matching the test 
prefix, as well as explicitly patching the `setup` method.